Add files via upload
I created a web app where a user is allowed to enter his/her name so he/she can be in conversation with Siri. The user enters 3 digits and the application check if it is a Palindrom or not. Then the user to view he/she must log in then the options of the summery and contiue are displayed.